modelTester/requirements.txt
huggingface_hub
datasets
kagglehub
pillow
torch

asyncssh
python-certifi-win32

pydantic
pydantic-core
pydantic-settings